# Conflicts:
#	Explorer/Assets/DCL/PerformanceAndDiagnostics/Diagnostics/ReportsHandling/ReportCategory.cs

+++ Explorer/Assets/Scripts/ECS/StreamableLoading/Cache/Disk/DiskCache.cs
@@ -1,5 +1,6 @@
 using Cysharp.Threading.Tasks;
 using ECS.StreamableLoading.Cache.Disk.CleanUp;
+using ECS.StreamableLoading.Cache.Disk.Lock;
 using System;
 using System.Buffers;
 using System.IO;
@@ -12,21 +13,28 @@ namespace ECS.StreamableLoading.Cache.Disk
     public class DiskCache : IDiskCache
     {
         private readonly CacheDirectory directory;
+        private readonly FilesLock filesLock;
         private readonly IDiskCleanUp diskCleanUp;
 
-        public DiskCache(CacheDirectory directory, IDiskCleanUp diskCleanUp)
+        public DiskCache(CacheDirectory directory, FilesLock filesLock, IDiskCleanUp diskCleanUp)
         {
             this.directory = directory;
+            this.filesLock = filesLock;
             this.diskCleanUp = diskCleanUp;
         }
 
-        public async UniTask<EnumResult<TaskError>> PutAsync(string key, string extension, ReadOnlyMemory<byte> data, CancellationToken token)
+        public async UniTask<EnumResult<TaskError>> PutAsync(HashKey key, string extension, ReadOnlyMemory<byte> data, CancellationToken token)
         {
             await using var scope = await ExecuteOnThreadPoolScope.NewScopeAsync();
 
             try
             {
                 string path = PathFrom(key, extension);
+                using var _ = filesLock.TryLock(path, out bool success);
+
+                if (success == false)
+                    return EnumResult<TaskError>.ErrorResult(TaskError.MessageError, "File is being used");
+
                 await using var stream = new FileStream(path, FileMode.Create, FileAccess.Write);
                 await stream.WriteAsync(data, token);
                 diskCleanUp.CleanUpIfNeeded();
@@ -38,7 +46,7 @@ namespace ECS.StreamableLoading.Cache.Disk
             return EnumResult<TaskError>.SuccessResult();
         }
 
-        public async UniTask<EnumResult<SlicedOwnedMemory<byte>?, TaskError>> ContentAsync(string key, string extension, CancellationToken token)
+        public async UniTask<EnumResult<SlicedOwnedMemory<byte>?, TaskError>> ContentAsync(HashKey key, string extension, CancellationToken token)
         {
             await using var scope = await ExecuteOnThreadPoolScope.NewScopeAsync();
 
@@ -46,6 +54,11 @@ namespace ECS.StreamableLoading.Cache.Disk
             {
                 string fileName = HashNamings.HashNameFrom(key, extension);
                 string path = PathFrom(fileName);
+
+                using var fileScope = filesLock.TryLock(path, out bool success);
+
+                if (success == false)
+                    return EnumResult<SlicedOwnedMemory<byte>?, TaskError>.SuccessResult(null);
 
                 if (File.Exists(path) == false)
                     return EnumResult<SlicedOwnedMemory<byte>?, TaskError>.SuccessResult(null);
@@ -62,20 +75,25 @@ namespace ECS.StreamableLoading.Cache.Disk
             catch (Exception e) { return EnumResult<SlicedOwnedMemory<byte>?, TaskError>.ErrorResult(TaskError.UnexpectedException, e.Message ?? string.Empty); }
         }
 
-        public async UniTask<EnumResult<TaskError>> RemoveAsync(string key, string extension, CancellationToken token)
+        public async UniTask<EnumResult<TaskError>> RemoveAsync(HashKey key, string extension, CancellationToken token)
         {
             await using var scope = await ExecuteOnThreadPoolScope.NewScopeAsync();
 
             try
             {
                 string path = PathFrom(key, extension);
+                using var fileScope = filesLock.TryLock(path, out bool success);
+
+                if (success == false)
+                    return EnumResult<TaskError>.ErrorResult(TaskError.MessageError, "File is being used");
+
                 if (File.Exists(path)) File.Delete(path);
                 return EnumResult<TaskError>.SuccessResult();
             }
             catch (Exception e) { return EnumResult<TaskError>.ErrorResult(TaskError.UnexpectedException, e.Message ?? string.Empty); }
         }
 
-        private string PathFrom(string key, string extension)
+        private string PathFrom(HashKey key, string extension)
         {
             string hashName = HashNamings.HashNameFrom(key, extension);
             return PathFrom(hashName);
@@ -99,13 +117,13 @@ namespace ECS.StreamableLoading.Cache.Disk
             this.serializer = serializer;
         }
 
-        public async UniTask<EnumResult<TaskError>> PutAsync(string key, string extension, T data, CancellationToken token)
+        public async UniTask<EnumResult<TaskError>> PutAsync(HashKey key, string extension, T data, CancellationToken token)
         {
             using SlicedOwnedMemory<byte> serializedData = await serializer.SerializeAsync(data, token);
             return await diskCache.PutAsync(key, extension, serializedData.Memory, token);
         }
 
-        public async UniTask<EnumResult<Option<T>, TaskError>> ContentAsync(string key, string extension, CancellationToken token)
+        public async UniTask<EnumResult<Option<T>, TaskError>> ContentAsync(HashKey key, string extension, CancellationToken token)
         {
             var result = await diskCache.ContentAsync(key, extension, token);
 
@@ -121,7 +139,7 @@ namespace ECS.StreamableLoading.Cache.Disk
             return EnumResult<Option<T>, TaskError>.SuccessResult(Option<T>.Some(deserializedValue));
         }
 
-        public UniTask<EnumResult<TaskError>> RemoveAsync(string key, string extension, CancellationToken token) =>
+        public UniTask<EnumResult<TaskError>> RemoveAsync(HashKey key, string extension, CancellationToken token) =>
             diskCache.RemoveAsync(key, extension, token);
     }
 }
